feat: MET-1253 modify start time and end time of epoch

+++ src/main/java/org/cardanofoundation/explorer/api/model/response/pool/DelegationHeaderResponse.java
@@ -16,7 +16,7 @@ public class DelegationHeaderResponse implements Serializable {
 
   private Long countDownEndTime;
 
-  private Long epochSlotNo;
+  private Integer epochSlotNo;
 
   private BigInteger liveStake;
